requirements.txt
ollama
faster-whisper
pyttsx3
pyaudio
pywin32
openpyxl
python-docx
azure-cognitiveservices-speech
PyQt5
PyQt5-tools
